fix(jitpack): remove 'v' prefix from tag
Remove the 'v' prefix from the tag since this causes additional builds
being caused by jitpack, since the release of 4.5.0 with a git tag of
v4.5.0 caused 2 builds to trigger.
jitpack.io reports v4.5.0 as the release, but both v4.5.0 and 4.5.0
are available in the 'builds' tab in jitpack. Better for it to only
have 4.5.0 available for disambiguation.